Rename the :sample module to :micro-apps
To disambiguate from ":sample-module", it was a bit confusing I think to
 have these two things but they are two quite different things.

Also move components to m3 package, to prepare for an m2 version of
them too.
