+++ .idea/runConfigurations/Design_Showcase.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Design Showcase" type="AndroidRunConfigurationType" factoryName="Android App">
-    <module name="hedvigandroid.sample.design-showcase.main" />
+    <module name="hedvigandroid.micro-apps.design-showcase.main" />
     <option name="DEPLOY" value="true" />
     <option name="DEPLOY_APK_FROM_BUNDLE" value="false" />
     <option name="DEPLOY_AS_INSTANT" value="false" />
