+++ micro-apps/README.md
@@ -1,0 +1,6 @@
+## micro-apps
+
+### Directory to be used to host micro-apps to quickly iterate on something
+
+Or for whatever else one might want to do with a light-weight version of the app, only depending on
+specific modules
