+++ micro-apps/design-showcase/src/main/AndroidManifest.xml
