+++ micro-apps/design-showcase/src/main/kotlin/com/hedvig/android/sample/design/showcase/ui/m3/components/Buttons.kt
@@ -17,39 +17,62 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.hedvig.android.sample.design.showcase.ui.components
+package com.hedvig.android.sample.design.showcase.ui.m3.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Checkbox
+import androidx.compose.material3.Button
+import androidx.compose.material3.ExtendedFloatingActionButton
+import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 
 @Composable
-internal fun M3Checkbox() {
-  var checked by remember { mutableStateOf(true) }
+internal fun M3Buttons() {
   Column {
+    val enabled by remember { mutableStateOf(true) }
     Spacer(Modifier.size(16.dp))
     M3OnSurfaceText(
-      text = "Checkbox",
+      text = "Buttons",
       style = MaterialTheme.typography.headlineSmall,
     )
     Spacer(Modifier.size(16.dp))
-    Checkbox(
-      checked = checked,
-      onCheckedChange = { checked = it },
-    )
-    Spacer(Modifier.size(8.dp))
-    Checkbox(
-      checked = !checked,
-      onCheckedChange = { checked = !it },
-    )
+    Button(
+      onClick = {},
+      enabled = enabled,
+    ) {
+      Text("Primary button")
+    }
+    Spacer(Modifier.size(16.dp))
+    OutlinedButton(
+      onClick = {},
+      enabled = enabled,
+    ) {
+      Text("Secondary button")
+    }
+    Spacer(Modifier.size(16.dp))
+    TextButton(
+      onClick = {},
+      enabled = enabled,
+    ) {
+      Text("Text button")
+    }
+    Spacer(Modifier.size(16.dp))
+    FloatingActionButton(onClick = {}) {
+      Text("FAB")
+    }
+    Spacer(Modifier.size(16.dp))
+    ExtendedFloatingActionButton(onClick = {}) {
+      Text("Extended FAB")
+    }
   }
 }
