+++ micro-apps/design-showcase/src/main/kotlin/com/hedvig/android/sample/design/showcase/ui/m3/components/Checkbox.kt
@@ -17,14 +17,13 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.hedvig.android.sample.design.showcase.ui.components
+package com.hedvig.android.sample.design.showcase.ui.m3.components
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Checkbox
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Switch
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -34,19 +33,23 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 
 @Composable
-internal fun M3Switch() {
+internal fun M3Checkbox() {
+  var checked by remember { mutableStateOf(true) }
   Column {
     Spacer(Modifier.size(16.dp))
     M3OnSurfaceText(
-      text = "Switch",
+      text = "Checkbox",
       style = MaterialTheme.typography.headlineSmall,
     )
     Spacer(Modifier.size(16.dp))
-    Row {
-      var checked by remember { mutableStateOf(false) }
-      Switch(checked = checked, onCheckedChange = { checked = !checked })
-      Spacer(modifier = Modifier.size(16.dp))
-      Switch(checked = !checked, onCheckedChange = { checked = !checked })
-    }
+    Checkbox(
+      checked = checked,
+      onCheckedChange = { checked = it },
+    )
+    Spacer(Modifier.size(8.dp))
+    Checkbox(
+      checked = !checked,
+      onCheckedChange = { checked = !it },
+    )
   }
 }
