+++ micro-apps/design-showcase/src/main/kotlin/com/hedvig/android/sample/design/showcase/ui/m3/components/Chips.kt
@@ -17,29 +17,50 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.hedvig.android.sample.design.showcase.ui.components
+package com.hedvig.android.sample.design.showcase.ui.m3.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.FilterChip
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.rememberDatePickerState
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.hedvig.android.core.designsystem.component.datepicker.HedvigDatePicker
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-internal fun M3DatePicker() {
+internal fun M3Chips() {
   Column {
     Spacer(Modifier.size(16.dp))
     M3OnSurfaceText(
-      text = "DatePicker",
+      text = "Chips",
       style = MaterialTheme.typography.headlineSmall,
     )
     Spacer(Modifier.size(16.dp))
-    HedvigDatePicker(rememberDatePickerState(), Modifier.size(500.dp))
+    var selectedId by remember { mutableStateOf(1) }
+    FilterChip(
+      label = { Text("Android") },
+      onClick = { selectedId = 1 },
+      selected = selectedId == 1,
+    )
+    Spacer(modifier = Modifier.size(4.dp))
+    FilterChip(
+      label = { Text("Material") },
+      onClick = { selectedId = 2 },
+      selected = selectedId == 2,
+    )
+    Spacer(modifier = Modifier.size(4.dp))
+    FilterChip(
+      label = { Text("Compose") },
+      onClick = { selectedId = 3 },
+      selected = selectedId == 3,
+    )
   }
 }
