+++ micro-apps/design-showcase/src/main/kotlin/com/hedvig/android/sample/design/showcase/ui/m3/components/ProgressBar.kt
@@ -17,42 +17,26 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.hedvig.android.sample.design.showcase.ui.components
+package com.hedvig.android.sample.design.showcase.ui.m3.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Tab
-import androidx.compose.material3.TabRow
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 
 @Composable
-internal fun M3Tab() {
-  var selectedIndex by remember { mutableStateOf(0) }
-  val tabs = listOf("Accounts", "Cards", "Funds")
+internal fun M3ProgressBar() {
   Column {
     Spacer(Modifier.size(16.dp))
     M3OnSurfaceText(
-      text = "Tab",
+      text = "Progress bar ",
       style = MaterialTheme.typography.headlineSmall,
     )
     Spacer(Modifier.size(16.dp))
-    TabRow(selectedTabIndex = selectedIndex) {
-      tabs.forEachIndexed { index, title ->
-        Tab(
-          text = { Text(title) },
-          selected = index == selectedIndex,
-          onClick = { selectedIndex = index },
-        )
-      }
-    }
+    CircularProgressIndicator()
   }
 }
