+++ micro-apps/design-showcase/src/main/kotlin/com/hedvig/android/sample/design/showcase/ui/m3/components/Switch.kt
@@ -17,21 +17,36 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.hedvig.android.sample.design.showcase.ui.components
+package com.hedvig.android.sample.design.showcase.ui.m3.components
 
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
+import androidx.compose.material3.Switch
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.text.TextStyle
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 
 @Composable
-internal fun M3OnSurfaceText(
-  text: String,
-  style: TextStyle,
-) {
-  Text(
-    color = MaterialTheme.colorScheme.onSurface,
-    text = text,
-    style = style,
-  )
+internal fun M3Switch() {
+  Column {
+    Spacer(Modifier.size(16.dp))
+    M3OnSurfaceText(
+      text = "Switch",
+      style = MaterialTheme.typography.headlineSmall,
+    )
+    Spacer(Modifier.size(16.dp))
+    Row {
+      var checked by remember { mutableStateOf(false) }
+      Switch(checked = checked, onCheckedChange = { checked = !checked })
+      Spacer(modifier = Modifier.size(16.dp))
+      Switch(checked = !checked, onCheckedChange = { checked = !checked })
+    }
+  }
 }
