+++ micro-apps/design-showcase/src/main/kotlin/com/hedvig/android/sample/design/showcase/ui/m3/components/Tab.kt
@@ -17,62 +17,42 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.hedvig.android.sample.design.showcase.ui.components
+package com.hedvig.android.sample.design.showcase.ui.m3.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Button
-import androidx.compose.material3.ExtendedFloatingActionButton
-import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Tab
+import androidx.compose.material3.TabRow
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 
 @Composable
-internal fun M3Buttons() {
+internal fun M3Tab() {
+  var selectedIndex by remember { mutableStateOf(0) }
+  val tabs = listOf("Accounts", "Cards", "Funds")
   Column {
-    val enabled by remember { mutableStateOf(true) }
     Spacer(Modifier.size(16.dp))
     M3OnSurfaceText(
-      text = "Buttons",
+      text = "Tab",
       style = MaterialTheme.typography.headlineSmall,
     )
     Spacer(Modifier.size(16.dp))
-    Button(
-      onClick = {},
-      enabled = enabled,
-    ) {
-      Text("Primary button")
-    }
-    Spacer(Modifier.size(16.dp))
-    OutlinedButton(
-      onClick = {},
-      enabled = enabled,
-    ) {
-      Text("Secondary button")
-    }
-    Spacer(Modifier.size(16.dp))
-    TextButton(
-      onClick = {},
-      enabled = enabled,
-    ) {
-      Text("Text button")
-    }
-    Spacer(Modifier.size(16.dp))
-    FloatingActionButton(onClick = {}) {
-      Text("FAB")
-    }
-    Spacer(Modifier.size(16.dp))
-    ExtendedFloatingActionButton(onClick = {}) {
-      Text("Extended FAB")
+    TabRow(selectedTabIndex = selectedIndex) {
+      tabs.forEachIndexed { index, title ->
+        Tab(
+          text = { Text(title) },
+          selected = index == selectedIndex,
+          onClick = { selectedIndex = index },
+        )
+      }
     }
   }
 }
