+++ micro-apps/design-showcase/src/main/res/values/strings.xml
