+++ settings.gradle.kts
@@ -60,6 +60,6 @@ include(":hanalytics:hanalytics-feature-flags")
 include(":hanalytics:hanalytics-feature-flags-test")
 include(":hedvig-language")
 include(":hedvig-market")
+include(":micro-apps:design-showcase")
 include(":notification-badge-data")
-include(":sample:design-showcase")
 include(":testdata")
